Rename sync agent module to agent
